+++ doc/changelog.md
@@ -54,7 +54,7 @@ Syntax: `- short text describing the change _(Your Name)_`
 - _()_
 - Improve loading spinner size in file selector and improve file selector responsiveness #1026 _(tschawax)_
 - _()_
-- _()_
+- Enable planting a field of a plant #1039 _(tschawax)_
 - _()_
 - _()_
 - _()_